requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.2
pyaudio==0.2.13
sounddevice==0.4.6
soundfile==0.12.1
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2 